+++ backend/database/index.js
@@ -6,5 +6,6 @@ const CLUSTER = 'fullstack-review';
 const uri = `mongodb+srv://Heroku:${PASSWORD}@fullstack-review.maljn.mongodb.net/ocean?retryWrites=true&w=majority`;
 
 const mongoose = require('mongoose');
-mongoose.connect(uri);
+mongoose.connect(uri, { useNewUrlParser: true,
+useUnifiedTopology: true });
 module.exports = mongoose;
