+++ backend/server/app.js
@@ -3,18 +3,14 @@ const express = require('express');
 const db = require('../database');
 const Models = require('../database/models.js');
 const app = express();
-const port = process.env.PORT || 3000;
+const port = 3000;
 
-// middleware
-// app.use(cors());
 app.use(express.json());
 app.use(express.static(__dirname + '/../../dist'));
-
 /*
   While writing your endpoints, please do not remove any of the mongo created
   _id properties
 **/
-
 app.get('/test', (req, res) => {
   // Models.User.add({userId: 0, name: 'Tom Marvolo Riddle'});
 });
