+++ webpack.config.js
@@ -7,6 +7,9 @@ const config = {
     path: path.resolve(__dirname, 'dist'),
     filename: 'bundle.js'
   },
+  devServer: {
+    historyApiFallback: true,
+  },
   module: {
     rules: [
       {
